requirements.txt
streamlit
langchain 
google-generativeai 
faiss-cpu 
python-dotenv
langchain.agents
serpapi
numexpr
google-search-results